Setting executable flag on klr which is lost on File.Copy
Fixes : https://github.com/aspnet/XRE/issues/1130
